final/misc/Face Unlock/setup/requirements.txt
gunicorn
flask
face_recognition
pillow
numpy